#637: Move season creation dialog to own directory
